Add import command to ymir
